Moved flatten function to the lib module.
